+++ src/components/Icon/view/IconsMap.tsx
@@ -157,6 +157,9 @@ export const IconsMap = {
     searchNoResults: <Icons.SearchNoResults />,
     userTimes: <Icons.UserTimes />,
     lockAltSolid: <Icons.LockAltSolid />,
+    dragAndDrop: <Icons.DragAndDrop />,
+    starSlash: <Icons.StarSlash />,
+    folders: <Icons.Folders />,
 };
 
 export type IconName = keyof typeof IconsMap;