+++ src/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './wrapCurried_DEPRECATED';
 export * from './wrap';
 export * from './handleParentRef';